Estimation or prediction for rrenv

Perform estimation or prediction under the reduced rank envelope model.

Usage

pred_eenvlp(m, Xnew)

Arguments

m

A list containing estimators and other statistics inherited from eenvlp.

Xnew

The value of X with which to estimate or predict Y. A p dimensional vector or an n by p matrix.

Details

This function evaluates the reduced rank envelope model at new value Xnew. It can perform estimation: find the fitted value when X = Xnew, or prediction: predict Y when X = Xnew. The covariance matrix and the standard errors are also provided.

Value

value

The fitted value or the predicted value evaluated at Xnew.

Examples

data(cereal)
m <- eenvlp(cereal$x[-1,], cereal$y[-1,], u=1, lamb=1e-07)
pred <- pred_eenvlp(m, cereal$x[1,])
pred
#>         [,1]     [,2]
#> [1,] -1.0287 1.350548